chore: refactor 'info' page rfc loading
